+++ src/blant.h
@@ -4,6 +4,7 @@
 #include <stdbool.h>
 #include "blant-fundamentals.h" // defining k related constants, including SELF_LOOPS
 #include "misc.h"
+#include "blant-pthreads.h"
 #include "sets.h"
 #include "tinygraph.h"
 #include "blant-window.h"
@@ -167,44 +168,9 @@ extern int _quiet; // suppress notes and/or warnings, higher value = more quiet
 
 Boolean NodeSetSeenRecently(GRAPH *G, unsigned Varray[], int k);
 
-// Headers for multithreading
-extern int _numThreads;
-extern int _maxThreads;
 // keeps track of if the stop mode is specified with number of samples (-n) or precision (-p)
-enum StopMode {num_samples, precision};
+enum StopMode {stopOnSamples, stopOnPrecision};
 extern enum StopMode _stopMode;
-
-typedef struct {
-    // local accumulator values, they function the same as globals but ARE LOCAL TO THREADS
-    double graphletCount[MAX_CANONICALS];
-    double graphletConcentration[MAX_CANONICALS];
-    double *graphletDegreeVector[MAX_CANONICALS];
-    double *orbitDegreeVector[MAX_ORBITS];
-    SET*** communityNeighbors;
-} Accumulators;
-
-Accumulators* InitializeAccumulatorStruct(GRAPH* G);
-void FreeAccumulatorStruct(Accumulators *accums);
-
-// a global array of accumulators, one for each thread, updated in batches
-extern Accumulators *_threadAccumulators;
-
-// Anytime a function must take an Accumulator as a parameter, but you don't intend on actually using the data, pass it this. 
-// The data here is never used, and maintaining only one copy of this saves memory.
-extern Accumulators _trashAccumulator;
-
-// https://docs.oracle.com/cd/E19120-01/open.solaris/816-5137/tlib-4/index.html
-typedef struct {
-    int samplesPerThread;
-    int k;
-    GRAPH *G;
-    int varraySize;
-    int threadId; // thread number, starting from 0
-    long seed;
-    Accumulators *accums;
-} ThreadData;
-
-void SampleNGraphletsInThreads(int k, GRAPH *G, int varraySize, int numSamples, int numThreads);
 
 void* RunBlantInThread(void* arg);
